pyproject.toml
[project]
name = "fpv-metadata-scan"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "oss2>=2.19.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "anyio>=4.12.0",
    "certifi>=2026.1.4",
    "charset-normalizer>=3.4.4",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "idna>=3.11",
    "lark-oapi>=1.5.2",
    "pycryptodome>=3.23.0",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "typing-extensions>=4.15.0",
    "urllib3>=2.6.2",
    "websockets>=15.0.1",
]